refine T5 model
